+++ presentation/src/main/java/com/example/presentation/view/gallery/model/GalleryImage.kt
@@ -1,0 +1,8 @@
+package com.example.presentation.view.gallery.model
+
+import android.net.Uri
+
+data class GalleryImage(
+    val imageUri: Uri,
+    val check: Boolean = false,
+)
