+++ presentation/src/main/java/com/example/presentation/view/gallery/view/SelectImageFragment.kt
@@ -1,0 +1,10 @@
+package com.example.presentation.view.gallery.view
+
+import com.example.presentation.R
+import com.example.presentation.base.BaseFragment
+import com.example.presentation.databinding.FragmentGallerySelectImageBinding
+
+class SelectImageFragment : BaseFragment<FragmentGallerySelectImageBinding>(R.layout.fragment_gallery_select_image){
+    override fun init() {
+    }
+}
