+++ presentation/src/main/java/com/example/presentation/view/gallery/viewModel/GalleryViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.presentation.view.gallery.viewModel
+
+import androidx.lifecycle.ViewModel
+
+class GalleryViewModel: ViewModel() {
+}
