+++ presentation/src/main/res/layout/fragment_gallery_list.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_height="match_parent"
+        android:layout_width="match_parent">
+        <Toolbar
+            android:id="@+id/tb_gallery"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:contentInsetStart="0dp">
+            <ImageButton
+                android:layout_width="@dimen/global_button_size"
+                android:layout_height="@dimen/global_button_size"
+                android:background="@android:color/transparent"
+                android:src="@drawable/ic_arrow"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/gallery_all_photo"
+                android:layout_gravity="center"
+                style="@style/title_bold"
+                android:textColor="@color/color_black"/>
+
+            <Button
+                android:layout_width="@dimen/global_button_size"
+                android:layout_height="@dimen/global_button_size"
+                android:layout_gravity="end"
+                android:background="@android:color/transparent"
+                android:text="다음"
+                android:padding="0dp"/>
+        </Toolbar>
+        <!--        <com.takusemba.cropme.CropLayout-->
+        <!--            android:id="@+id/cl_cropper"-->
+        <!--            android:layout_width="match_parent"-->
+        <!--            android:layout_height="0dp"-->
+        <!--            app:cropme_background_alpha="80%"-->
+        <!--            app:cropme_frame_height_percent="80%"-->
+        <!--            app:cropme_frame_width_percent="80%"-->
+        <!--            app:layout_constraintDimensionRatio="1:1"-->
+        <!--            app:cropme_with_border="true"-->
+        <!--            app:cropme_max_scale="4.0"-->
+        <!--            app:cropme_overlay_shape="rectangle"-->
+        <!--            app:layout_constraintTop_toTopOf="parent"-->
+        <!--            app:layout_constraintEnd_toEndOf="parent"-->
+        <!--            app:layout_constraintStart_toStartOf="parent"/>-->
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_gallery"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintTop_toBottomOf="@id/tb_gallery"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+            app:spanCount="3"
+            tools:listitem="@layout/item_gallery"/>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
